# application/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于区块链技术的房地产交易系统api文档",
        "title": "基于区块链技术的房地产交易系统api文档",
        "contact": {
            "name": "togettoyou",
            "email": "zoujh99@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/createRealEstate": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "新建房地产(管理员)",
                "parameters": [
                    {
                        "description": "realEstate",
                        "name": "realEstate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.RealEstateRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/queryRealEstateList": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取房地产信息(空json{}可以查询所有，指定proprietor可以查询指定业主名下商品)",
                "parameters": [
                    {
                        "description": "realEstateQuery",
                        "name": "realEstateQuery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.RealEstateQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "v1.RealEstateQueryRequestBody": {
            "type": "object",
            "properties": {
                "proprietor": {
                    "description": "所有者(业主)(业主AccountId)",
                    "type": "string"
                }
            }
        },
        "v1.RealEstateRequestBody": {
            "type": "object",
            "properties": {
                "accountId": {
                    "description": "操作人ID",
                    "type": "string"
                },
                "livingSpace": {
                    "description": "新旧程度",
                    "type": "number"
                },
                "proprietor": {
                    "description": "所有者(业主)(业主AccountId)",
                    "type": "string"
                },
                "totalArea": {
                    "description": "购买时间",
                    "type": "number"
                }
            }
        },
...
    }
}